+++ opentelemetry-sdk/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+- Rename CounterAggregator -> SumAggregator
+  ([#816](https://github.com/open-telemetry/opentelemetry-python/pull/816))
+
+## 0.9b0
+
+Released 2020-06-10
+
 - Move stateful & resource from Meter to MeterProvider
   ([#751](https://github.com/open-telemetry/opentelemetry-python/pull/751))
 - Rename Measure to ValueRecorder in metrics